requirements.txt
Flask==2.3.3
dashscope==1.14.0
flask-cors==4.0.0
python-dotenv==1.0.0
Pillow==10.0.1
requests==2.31.0
pytesseract==0.3.10
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
